refactor: udpate file structure - follow latest practices
