+++ src/App.tsx
@@ -1,0 +1,9 @@
+import { FC } from 'react';
+
+import { AppProps } from './lib/types/app';
+
+import { BuyDomain } from './pages';
+
+export const App: FC<AppProps> = () => {
+	return <BuyDomain />;
+};
