--- src/components/index.ts
@@ -1,5 +1,0 @@
-export * from './BuyDomainModal';
-export * from './DomainSearch';
-export * from './QuestionButton';
-export * from './ConnectWallet';
-export * from './Icons';
