+++ src/features/buy-domain/BuyDomainModal/BuyDomainModal.module.scss
@@ -1,6 +1,6 @@
-@import '../../../node_modules/@zero-tech/zui/styles/_colors.scss';
-@import '../../../node_modules/@zero-tech/zui/styles/_dimensions.scss';
-@import '../../../node_modules/@zero-tech/zui/styles/_typography.scss';
+@import '../../../../node_modules/@zero-tech/zui/styles/_colors.scss';
+@import '../../../../node_modules/@zero-tech/zui/styles/_dimensions.scss';
+@import '../../../../node_modules/@zero-tech/zui/styles/_typography.scss';
 
 .Modal {
 	padding: unset;
