+++ src/features/buy-domain/BuyDomainModal/Elements/index.ts
@@ -1,3 +1,4 @@
+export * from './DefaultDomainMedia';
 export * from './Discover';
 export * from './Approve';
 export * from './Mint';
