+++ src/features/buy-domain/BuyDomainModal/index.ts
@@ -1,0 +1,2 @@
+export * from './BuyDomainModal';
+export * from './Elements';
