+++ src/features/buy-domain/Icons/IconClose.tsx
