+++ src/features/buy-domain/Icons/IconZero.tsx
