+++ src/features/buy-domain/Icons/index.ts
