+++ src/features/buy-domain/index.ts
@@ -1,1 +1,2 @@
 export * from './BuyDomainModal';
+export * from './Icons';
