+++ src/features/domain-search/DomainSearch.module.scss
