+++ src/features/ui/QuestionButton/QuestionButton.module.scss
@@ -1,6 +1,6 @@
-@import '../../../node_modules/@zero-tech//zui/styles/_animations.scss';
-@import '../../../node_modules/@zero-tech/zui/styles/_borders.scss';
-@import '../../../node_modules/@zero-tech/zui/styles/_colors.scss';
+@import '../../../../node_modules/@zero-tech//zui/styles/_animations.scss';
+@import '../../../../node_modules/@zero-tech/zui/styles/_borders.scss';
+@import '../../../../node_modules/@zero-tech/zui/styles/_colors.scss';
 
 .QuestionButton {
 	position: relative;
