+++ src/lib/constants/domains.ts
