+++ src/lib/constants/network.ts
