+++ src/lib/helpers/network.ts
