+++ src/lib/hooks/useDomainAvailability.ts
