+++ src/lib/util/ChainGate.tsx
@@ -1,6 +1,6 @@
+import React from 'react';
 import type { FC, ReactNode } from 'react';
 
-import React from 'react';
 import { Network } from '../constants/network';
 import { useWeb3 } from '../hooks';
