--- src/types.ts
@@ -1,11 +1,0 @@
-import { ethers } from 'ethers';
-
-export interface AppProps {
-	provider: any;
-	route: string;
-	web3: {
-		chainId: number;
-		address: string;
-		connectWallet;
-	};
-}
